chore: moving customization package to embed package
move the logic from the customization package into the embed package.
